Very important code change !!!!!
Math is a torturous subject that should be considered as child abuse to the poor school children that are forced to endure it. If you like math, stop. You are the reason that so many kids don't want to try anymore. Math isn't real. Stop spreading lies, it isn't funny.